requirements.txt
streamlit
pandas
numpy
propy3